+++ src/endpoints/postEstudante.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
 import EstudantesDataBase from "../data/EstudantesDataBase";
-import UsuarioModel from "../model/UsuarioModel";
+import UsuarioModel from "../model/types";
 
 async function postEstudante(req: Request, res: Response): Promise<void> {
     try {
